Twitter may be getting injected; able to fetch timeline informations.
