+++ src/main/java/liuni/dagger/TwitterComponent.java
@@ -1,0 +1,9 @@
+package liuni.dagger;
+
+import dagger.Component;
+import twitter4j.Twitter;
+
+@Component(modules = TwitterInjectionModule.class)
+public interface TwitterComponent {
+    Twitter injectTwitter();
+}
